sample car CRUD operation

+++ carRental/src/main/java/edu/miu/carRental/repository/CarRepository.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import edu.miu.carRental.domain.Car;
 @Repository
-public interface CarRepository extends JpaRepository<Car,Integer>{
+public interface CarRepository extends JpaRepository<Car,Long>{
+	
 
 }
